add session class
