+++ src/main/java/main/RestApplication.java
@@ -1,6 +1,7 @@
 package main;
 
 import rest.Users;
+import rest.Session;
 
 import javax.ws.rs.ApplicationPath;
 import javax.ws.rs.core.Application;
@@ -15,7 +16,9 @@ public class RestApplication extends Application {
     @Override
     public Set<Object> getSingletons() {
         final HashSet<Object> objects = new HashSet<>();
-        objects.add(new Users(new AccountService()));
+        AccountService accountService = new AccountService();
+        objects.add(new Session(accountService));
+        objects.add(new Users(accountService));
         return objects;
     }
 }
